EmployeeTest class
Here, I have created three Employee objects. These objects are then stored in an array named projectGroup. Then, I created a loop to iterate through each employee in the projectGroup array. The loop checks whether the employee's identification number condition represented by the variable m is true. If this condition is met, the program prints the name of the employee