Menu & EmployeeTest
Here, I incorporate a new manager and company to be passed as arguments to my system and be used for testing. I also created code to add employees to the ArrayList, which will be used in other parts of the code. Additionally, I implemented a code to add the employees to the company for functionality testing and a `startMenu` that will initiate the manager's menu, and add the employ number on display to be easier in case the manager need to locate the employee number to remove it from the system.

I designed the menu's code using a switch to get the user's choice while navigating the menu. The code includes methods to view the current staff in the system and methods to add and remove staff based on their employee numbers.